+++ v3/crates/engine/src/routes.rs
@@ -24,9 +24,7 @@ use crate::{
 
 use super::types::RequestType;
 
-const MB: usize = 1_048_576;
-
-pub fn get_base_routes(state: EngineState) -> Router {
+pub fn get_base_routes(state: EngineState, request_body_limit: usize) -> Router {
     let graphql_ws_route = Router::new()
         .route("/graphql", get(handle_websocket_request))
         .layer(axum::middleware::from_fn(|request, next| {
@@ -104,8 +102,8 @@ pub fn get_base_routes(state: EngineState) -> Router {
         .merge(health_route)
         // The '/*path' route
         .merge(pre_route_router)
-        // Set request payload limit to 10 MB
-        .layer(DefaultBodyLimit::max(10 * MB))
+        // Set request payload limit
+        .layer(DefaultBodyLimit::max(request_body_limit))
 }
 
 /// Serve the introspection metadata file and its hash at `/metadata` and `/metadata-hash` respectively.